moved html format to a sub directory
